requirements.txt
fastapi
uvicorn
websockets
llama-cpp-python
sqlalchemy